requirements.txt
Flask==3.0.3
Pillow==10.3.0
reportlab==4.1.0
PyInstaller==6.9.0
